tsak5

+++ src/services/deleteContact.js
@@ -1,8 +1,8 @@
 import Contacts from "../db/models/Contacts.js"
 
-const deleteContact = async (contactId) => {
+const deleteContact = async (contactId, userId) => {
     // throw new Error ('no result found')
-    const noContact = await Contacts.findOneAndDelete({_id: contactId});
+    const noContact = await Contacts.findOneAndDelete({_id: contactId, userId});
     return noContact;
     }
